refactor(maps): move property map to new component
added env variable for api key
