+++ frontend/.env.example
@@ -1,0 +1,1 @@
+REACT_APP_GOOGLE_API_KEY=
