+++ frontend/src/components/maps/propertyMapStyles.json
